+++ next/src/components/nomenclatures/NomenclatureOverviewItem.tsx
@@ -1,7 +1,7 @@
 import OverviewItem from '@/components/ui/OverviewItem'
 import type { Nomenclature } from '@/models/nomenclature'
 
-import RelatedQuestions from '../ui/RelatedQuestions'
+import NomenclatureOverviewItemContent from './NomenclatureOverviewItemContent'
 
 interface NomenclatureOverviewItemProps {
   nomenclature: Nomenclature
@@ -13,19 +13,7 @@ export default function NomenclatureOverviewItem({
 }: Readonly<NomenclatureOverviewItemProps>) {
   return (
     <OverviewItem
-      content={
-        <div className="flex justify-between items-center">
-          <div className="flex flex-col">
-            <h3>{nomenclature.label}</h3>
-            <span className="text-sm text-gray-600">
-              {nomenclature.version}
-            </span>
-          </div>
-          <RelatedQuestions
-            relatedQuestionNames={nomenclature.relatedQuestionNames}
-          />
-        </div>
-      }
+      content={<NomenclatureOverviewItemContent nomenclature={nomenclature} />}
     />
   )
 }