+++ next/src/models/nomenclature.ts
@@ -1,11 +1,8 @@
 export type Nomenclature = {
   id: string
   label: string
-  version: string
-  externalLink: ExternalLink
   relatedQuestionNames: string[]
-}
-
-type ExternalLink = {
-  urn: string
+  version?: string
+  referenceYear?: string
+  theme?: string
 }