+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Nomenclature page now displays the nomenclature theme, reference year and version.
+
 ## [3.4.1](https://github.com/InseeFr/Pogues/releases/tag/3.4.1) - 2026-07-03
 
 ## Fixed